requirements.txt
openai>=0.27.0
prompt_toolkit>=3.0.0
PyYAML>=5.0.0
pydantic
